backend/pyproject.toml
[tool.poetry]
name = "finalproject"
version = "0.2.0"
description = ""
authors = ["sangwoonoel <sangwoonoel@cau.ac.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.105.0"}
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
httpx = "^0.27.0"
pytest = "^8.0.2"
pymongo = "^4.6.2"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
commitizen = "^3.18.0"
pre-commit = "^3.6.2"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
